+++ asyncz/executors/pool.py
@@ -1,4 +1,5 @@
 import concurrent.futures
+import sys
 from abc import abstractmethod
 from concurrent.futures.process import BrokenProcessPool
 from datetime import datetime
@@ -14,6 +15,8 @@ if TYPE_CHECKING:
 
 class BasePoolExecutor(BaseExecutor):
     model_config = ConfigDict(extra="allow", arbitrary_types_allowed=True, populate_by_name=True)
+    cancel_futures: bool = False
+    overwrite_wait: Optional[bool] = None
 
     @abstractmethod
     def __init__(self, pool: Any, **kwargs: Any) -> None:
@@ -21,6 +24,9 @@ class BasePoolExecutor(BaseExecutor):
         self.pool = pool
 
     def do_send_task(self, task: "TaskType", run_times: List[datetime]) -> Any:
+        task_id = task.id
+        assert task_id is not None, "Cannot send decorator type task"
+
         def callback(fn: Any) -> None:
             exc, _ = (
                 fn.exception_info()
@@ -28,9 +34,9 @@ class BasePoolExecutor(BaseExecutor):
                 else (fn.exception(), getattr(fn.exception(), "__traceback__", None))
             )
             if exc:
-                self.run_task_error(task.id)
+                self.run_task_error(task_id)
             else:
-                self.run_task_success(task.id, fn.result())
+                self.run_task_success(task_id, fn.result())
 
         try:
             fn = self.pool.submit(run_task, task, task.store_alias, run_times)
@@ -42,6 +48,10 @@ class BasePoolExecutor(BaseExecutor):
         fn.add_done_callback(callback)
 
     def shutdown(self, wait: bool = True) -> None:
+        if self.overwrite_wait is not None:
+            wait = self.overwrite_wait
+        if sys.version_info >= (3, 9):
+            self.pool.shutdown(wait, cancel_futures=self.cancel_futures)
         self.pool.shutdown(wait)
 
 
@@ -54,10 +64,10 @@ class ThreadPoolExecutor(BasePoolExecutor):
         pool_kwargs: Dict of keyword arguments to pass to the underlying ThreadPoolExecutor constructor.
     """
 
-    def __init__(self, max_workers: int = 10, pool_kwargs: Optional[Any] = None):
+    def __init__(self, max_workers: int = 10, pool_kwargs: Optional[Any] = None, **kwargs: Any):
         pool_kwargs = pool_kwargs or {}
         pool = concurrent.futures.ThreadPoolExecutor(int(max_workers), **pool_kwargs)
-        super().__init__(pool)
+        super().__init__(pool, **kwargs)
 
 
 class ProcessPoolExecutor(BasePoolExecutor):
@@ -70,7 +80,9 @@ class ProcessPoolExecutor(BasePoolExecutor):
             ProcessPoolExecutor constructor.
     """
 
-    def __init__(self, max_workers: int = 10, pool_kwargs: Optional[Any] = None) -> None:
+    def __init__(
+        self, max_workers: int = 10, pool_kwargs: Optional[Any] = None, **kwargs: Any
+    ) -> None:
         pool_kwargs = pool_kwargs or {}
         pool = concurrent.futures.ProcessPoolExecutor(int(max_workers), **pool_kwargs)
-        super().__init__(pool)
+        super().__init__(pool, **kwargs)