+++ docs_src/triggers/date/example2.py
@@ -10,6 +10,6 @@ def my_task(number):
 
 
 # Execute the task on December 25th, 2022
-scheduler.add_task(my_task, run_date="2022-12-25 00:01:00", args=[25])
+scheduler.add_task(my_task, run_at="2022-12-25 00:01:00", args=[25])
 
 scheduler.start()